+++ src/main/java/com/yuuko/core/metrics/handlers/MetricsManager.java
@@ -1,13 +1,9 @@
 package com.yuuko.core.metrics.handlers;
 
-import com.yuuko.core.Configuration;
 import com.yuuko.core.metrics.DatabaseMetrics;
 import com.yuuko.core.metrics.DiscordMetrics;
 import com.yuuko.core.metrics.EventMetrics;
 import com.yuuko.core.metrics.SystemMetrics;
-import net.dv8tion.jda.core.entities.Guild;
-
-import java.util.List;
 
 public class MetricsManager {
 
@@ -37,25 +33,4 @@ public class MetricsManager {
         return systemMetrics;
     }
 
-    public static void updateDiscordMetrics() {
-        List<Guild> guilds = Configuration.BOT.getJDA().getGuilds();
-
-        int userCount = 0;
-        int channelCount = 0;
-        int emoteCount = 0;
-        int roleCount = 0;
-
-        for(Guild guild : guilds) {
-            userCount += guild.getMemberCache().size();
-            channelCount += guild.getChannels().size();
-            emoteCount += guild.getEmoteCache().size();
-            roleCount += guild.getRoleCache().size();
-        }
-
-        discordMetrics.GUILD_COUNT = guilds.size();
-        discordMetrics.USER_COUNT = userCount;
-        discordMetrics.CHANNEL_COUNT = channelCount;
-        discordMetrics.EMOTE_COUNT = emoteCount;
-        discordMetrics.ROLE_COUNT = roleCount;
-    }
 }