+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
+## 20-03-2019_1
++ [~] Updated Event Metrics to include more specific data about which types of user is behind each event, also added messages sent metrics.
+
 ## 19-03-2019_1
 + [~] Removed listening for numeric values between 1-10 and the word cancel in GenericMessageController and refactored the class.
 + [~] Altered how [__search__] works, you now need to use -search <value> or -search cancel to make the input less confusing.